CHANGES TO FIT SEVERAL ECHANS TOGETHER WITH ONE cgb_spectrum and one earth spectrum:
Continuous_data:
change that the method cgb_rate_array return the cgb_rate_array for all Echans. Which one is used is decided later in the source object.

SOURCE:
New tag for all sources, which states for which Echan this source is used

Added new source Class 'GlobalSource' which will be all sources which are connected over all echans

SETUP_SOURCE:
Added the new echan tag to all source creations (not to the global)

New method setup_sources_global: A method that creates the global sources that are needed.

MODEL:

changed everything to fit for the new GlobalSources and rewrote the get_counts method to only get the counts which belong to a certain echan

FUNCTIONS:
changed the class definitions to include the new GlobalFunction and renamed the different Sources such that one knows which source is for which echan

FUNCTION:
rewrote the call method to get an echan value as input
wrote a new Function class GlobalFunction which uses this new input value to choose the write function array, all other classes ignore this new input value

BACKGROUND_LIKE:
rewrote a lot:

1. now no echan is given but a List of echans.
2. The counts are imported for all echans
3. The evaluate_model method now only saves the model_counts from sources with the right echan tag
4. The call method now calculates the log_likelihood for every echan seperatly and sums them together in the end
5. For this a new method _get_log_likelihood_echan is written, that returns the log_likelihodd for a certain echan
6. Display model method has been adjusted to fit the new methods and plot the counts and model for a certain echan
7. In get_list_of_sources all sources that do not contribute to the current echan are dismissed and are not plotted