Included cspec. The fitting can be run now with ctime and cspec data.
Outsourced the download of the files from continousdata to a separate download method.

Fixed the rebin of the time bins before the fitting.

In rate_gernerator_DRM read in the current Ebins for the detected photons
for the used day from the data file of the day. Automatically chooses the right
Ebins for ctime and cspec. (For cspec the precalculation takes much longer, because
there are more Ebins to calculate the expected rate)